# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Timetable API",
        "description": "API documentation for the Timetable application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/generated-timetables": {
            "get": {
                "tags": [
                    "Generated Timetables"
                ],
                "summary": "Get the active generated timetable for the authenticated user",
                "description": "Returns the active generated timetable",
                "operationId": "getActiveGeneratedTimetable",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneratedTimetable"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "No active timetable found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Generated Timetables"
                ],
                "summary": "Generate a new timetable based on user preferences",
                "description": "Returns a newly generated timetable",
                "operationId": "generateTimetable",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneratedTimetable"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Failed to generate timetable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lecturers": {
            "get": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Get list of lecturers",
                "description": "Returns list of lecturers",
                "operationId": "getLecturersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Lecturer"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Store new lecturer",
                "description": "Returns lecturer data",
                "operationId": "storeLecturer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Lecturer"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lecturer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/lecturers/{id}": {
            "get": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Get lecturer information",
                "description": "Returns lecturer data",
                "operationId": "getLecturerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lecturer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lecturer"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Update existing lecturer",
                "description": "Returns updated lecturer data",
                "operationId": "updateLecturer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lecturer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Lecturer"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lecturer"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lecturers"
                ],
                "summary": "Delete existing lecturer",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteLecturer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lecturer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GeneratedTimetable": {
                "title": "Generated Timetable",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "user_id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "timetable": {
                        "type": "object"
                    },
                    "active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Lecturer": {
                "title": "Lecturer",
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john.doe@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "123-456-7890"
                    },
                    "department": {
                        "type": "string",
                        "example": "Computer Science"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}